Fixing Comments::delete_these() on two levels.
First, looks like chrisjdavis may have committed something he didn't mean to (in r3222) that broke it all together while working on optimizing the method. I'll let him come back with a working version of that later. By the way, using implode() would shrink that loop into a single line. :)

Second, the possibility that the method could receive: a Comments object, a single Comment object, a single comment ID, an array of Comment objects, or an array of Comment IDs didn't seem to be taken into account properly. I suspect I've missed some permutation of all that, but then we're really no worse off than we already were, right?

git-svn-id: http://svn.habariproject.org/habari/trunk/htdocs/system@3224 653ae4dd-d31e-0410-96ef-6bf7bf53c507

+++ classes/comments.php
@@ -324,33 +324,40 @@ class Comments extends ArrayObject
 		if ( ! is_array( $comments ) && ! $comments instanceOf Comments ) {
 			$comments = array( $comments );
 		}
-		
-		$in = '';
-		$i = 0;
-		foreach ( $comments as $comment ) {
-			if ( $i > 0 ) {
-				$in .= ', ' . $comment->id;
-			}
-			else {
-				$in .= $comment->id;
-			}
-		}
-		
-		echo $in; exit;
-		
+
 		if ( count( $comments ) == 0 ) {
 			return true;
 		}
 
-		if ( $comments[0] instanceOf Comment ) {
-			// We were passed an array of comment objects. Use them directly.
+		if ( $comments instanceOf Comments ) {
+			// Delete all the comments directly
 			$result = $comments->delete();
 		}
-		else if ( is_numeric( $comments[0] ) ) {
-			// We were passed an array of ID's. Get their objects and delete them.
-			$comments = self::get( array( 'id' => $comments ) );
+		else if ( is_array( $comments ) ) {
+			// We have an array... of something
 
-			$result = $comments->delete();
+			if ( $comments[0] instanceOf Comment ) {
+
+				$result = true;
+				foreach( $comments as $comment ) {
+					$comment_result = $comment->delete();
+
+					if ( !$comment_result ) {
+						$result = false;
+					}
+				}
+
+			}
+			else if ( is_numeric( $comments[0] ) ) {
+				// We were passed an array of ID's. Get their objects and delete them.
+				$comments = self::get( array( 'id' => $comments ) );
+
+				$result = $comments->delete();
+			}
+			else {
+				$result = false;
+			}
+
 		}
 		else {
 			// We were passed a type we could not understand.
